requirements.txt
flasgger==0.9.3
Flask==1.1.1
fuzzywuzzy==0.17.0
pandas
PyPDF2==1.26.0
python-docx==0.8.10
python-Levenshtein==0.12.0
regex==2019.11.1
requests==2.22.0
Rtree==0.8.3
scikit-learn
Wand==0.5.7
rq==1.1.0
redis==3.3.11
waitress==1.4.3
applicationinsights==0.11.9
Pillow==8.1.0
tensorflow-cpu==2.4.0
mtcnn==0.1.0
opencv-python==4.5.1.48
